INNOVA_PLUS_SANTE/requirements.txt
# INNOVA+ SANTÉ - Requirements
# Plateforme d'IA pour l'analyse prédictive de données de santé

# Interface utilisateur
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.4
streamlit-plotly-events>=0.0.6

# Data Science & ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Modèles prédictifs avancés
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.30.0

# Système RAG et LLM
langchain>=0.0.300
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=0.27.0
huggingface-hub>=0.16.0

# Base de données et stockage
sqlalchemy>=2.0.0
redis>=4.6.0

# APIs et services
requests>=2.31.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Utilitaires
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0
tqdm>=4.65.0
joblib>=1.3.0

# Tests et qualité
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.2.0

# Backend API
fastapi>=0.110.0
uvicorn[standard]>=0.24.0

# Data ingestion utilities
pyyaml>=6.0
kaggle>=1.6.14
pyarrow>=12.0.0
